% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiasedpathsampling-package.R
\docType{package}
\name{unbiasedpathsampling-package}
\alias{unbiasedpathsampling-package}
\alias{unbiasedpathsampling}
\title{unbiasedpathsampling}
\description{
Unbiased Path Sampling
}
\details{
This package contains scripts to remove the bias of path sampling estimators.
}
\author{
Pierre E. Jacob <pjacob@fas.harvard.edu>
}
\keyword{package}
